Replaced Invalid Cmdlets
Updated cmdlets: New-AzImageBuilderTemplateSourceObject, New-AzImageBuilderTemplateCustomizerObject and New-AzImageBuilderTemplateDistributorObject
New-AzImageBuilderTemplateSourceObject was prompting for PlatformImageSource and list of parameters changed.
New-AzImageBuilderTemplateCustomizerObject with PowerShellCustomizer $true wasn't working, it kept prompting for the value when running the script. Appears to require to be passed without a value.

Not changed but something I decided to use:
Invoke-Expression ". \Get-AzureImageInfo.ps1" instead of . E1\Get-AzureImageInfo.ps1